If we're using open_file, this actually belongs in test_smbclient_os.py.
